Refactor UI code into separate widget modules
Organize monitor UI code from a single ui.rs file into a ui module with separate files for each widget component (layout, status_bar, product, separator, hints, content).
