+++ rust/agama-cli/src/monitor/ui/content.rs
@@ -18,189 +18,19 @@
 // To contact SUSE LLC about this file by physical or electronic mail, you may
 // find current contact information at www.suse.com.
 
-//! UI rendering modules for the monitor TUI
+//! Content area widget for the monitor TUI
 
 use agama_lib::monitor::InstallationStatus;
-use agama_utils::api::{status::Stage, Issue, Scope};
+use agama_utils::api::{status::Stage, Scope};
 use gettextrs::gettext;
 use ratatui::{
     buffer::Buffer,
-    layout::{Constraint, Layout, Rect},
+    layout::Rect,
     style::{Color, Modifier, Style},
     text::{Line, Span},
-    widgets::{Gauge, List, Paragraph, Widget},
+    widgets::{Gauge, Paragraph, Widget},
 };
 use std::collections::HashMap;
-
-/// Layout areas for the monitor UI
-pub struct MonitorLayout {
-    /// Status bar (row 1)
-    pub status_bar: Rect,
-    /// Empty gap (row 2) - used for visual spacing
-    #[allow(dead_code)]
-    pub gap: Rect,
-    /// Product name (row 3)
-    pub product: Rect,
-    /// Separator line (row 4)
-    pub separator: Rect,
-    /// Content area (middle)
-    pub content: Rect,
-    /// Hints separator (bottom - 2)
-    pub hints_separator: Rect,
-    /// Hints footer (bottom - 1)
-    pub hints: Rect,
-}
-
-/// Creates the main layout for the monitor UI
-///
-/// Layout structure (matching TypeScript mockup + hints):
-/// - Row 1: Status bar (hostname, IP, machine, BUSY/IDLE, phase)
-/// - Row 2: Empty gap
-/// - Row 3: Product name
-/// - Row 4: Separator line
-/// - Middle: Dynamic content (progress, issues, messages) - with air gaps
-/// - Footer: Hints separator and keyboard hints (non-sticky, immediately below content)
-pub fn create_layout(area: Rect) -> MonitorLayout {
-    // Calculate content height: total - (status + gap + product + separator)
-    // Leave room for hints at bottom but don't make them sticky
-    let content_start = 4;
-    let content_height = area.height.saturating_sub(content_start);
-
-    let chunks = Layout::vertical([
-        Constraint::Length(1),              // Status bar
-        Constraint::Length(1),              // Gap
-        Constraint::Length(1),              // Product name
-        Constraint::Length(1),              // Separator
-        Constraint::Length(content_height), // Content + hints (non-sticky)
-    ])
-    .split(area);
-
-    // Split content area to have hints at bottom (but not screen-sticky)
-    let content_and_hints = Layout::vertical([
-        Constraint::Min(1),    // Content area (flexible)
-        Constraint::Length(1), // Hints separator
-        Constraint::Length(1), // Hints footer
-    ])
-    .split(chunks[4]);
-
-    MonitorLayout {
-        status_bar: chunks[0],
-        gap: chunks[1],
-        product: chunks[2],
-        separator: chunks[3],
-        content: content_and_hints[0],
-        hints_separator: content_and_hints[1],
-        hints: content_and_hints[2],
-    }
-}
-
-/// Renders the status bar
-pub fn render_status_bar(status: &InstallationStatus, area: Rect, buf: &mut Buffer) {
-    let width = area.width as usize;
-
-    // Determine busy state
-    let busy_state = if status.status.stage == Stage::Failed {
-        BusyState::Failed
-    } else if status.status.stage == Stage::Finished {
-        BusyState::Idle
-    } else if !status.status.progresses.is_empty() {
-        BusyState::Busy
-    } else if !status.issues.is_empty() || !status.questions.is_empty() {
-        BusyState::Waiting
-    } else {
-        BusyState::Idle
-    };
-
-    // Build left side: BUSY/IDLE badge + PHASE badge
-    let (busy_text, busy_color, busy_bg) = match busy_state {
-        BusyState::Busy => (" BUSY ", Color::Black, Color::Yellow),
-        BusyState::Waiting => (" IDLE ", Color::White, Color::Magenta),
-        BusyState::Failed => (" FAIL ", Color::White, Color::Red),
-        BusyState::Idle => (" IDLE ", Color::Black, Color::Green),
-    };
-
-    let (phase_text, phase_color, phase_bg) = match status.status.stage {
-        Stage::Installing => (" INSTALLING ", Color::Black, Color::Green),
-        Stage::Configuring => (" CONFIGURING ", Color::White, Color::Cyan),
-        Stage::Finished => (" FINISHED ", Color::Black, Color::Green),
-        Stage::Failed => (" FAILED ", Color::White, Color::Red),
-    };
-
-    // Build right side: hostname @ IP | machine (hide machine if unknown)
-    let right = if status.system_info.machine == "Unknown Machine" {
-        format!(
-            " {} @ {} ",
-            status.system_info.hostname, status.system_info.ip
-        )
-    } else {
-        format!(
-            " {} @ {} | {} ",
-            status.system_info.hostname, status.system_info.ip, status.system_info.machine
-        )
-    };
-
-    // Calculate gap
-    let left_plain_len = busy_text.len() + 1 + phase_text.len();
-    let gap = width.saturating_sub(left_plain_len + right.len());
-
-    // Build styled line
-    let line = Line::from(vec![
-        Span::styled(
-            busy_text,
-            Style::default()
-                .fg(busy_color)
-                .bg(busy_bg)
-                .add_modifier(Modifier::BOLD),
-        ),
-        Span::raw(" "),
-        Span::styled(
-            phase_text,
-            Style::default()
-                .fg(phase_color)
-                .bg(phase_bg)
-                .add_modifier(Modifier::BOLD),
-        ),
-        Span::styled(" ".repeat(gap), Style::default().bg(Color::DarkGray)),
-        Span::styled(
-            right,
-            Style::default()
-                .fg(Color::White)
-                .bg(Color::DarkGray)
-                .add_modifier(Modifier::BOLD),
-        ),
-    ]);
-
-    Paragraph::new(line)
-        .style(Style::default().bg(Color::DarkGray))
-        .render(area, buf);
-}
-
-/// Renders the product name
-pub fn render_product(product_name: &str, area: Rect, buf: &mut Buffer) {
-    Paragraph::new(format!(" {}", product_name))
-        .style(Style::default().add_modifier(Modifier::BOLD))
-        .render(area, buf);
-}
-
-/// Renders a separator line
-pub fn render_separator(area: Rect, buf: &mut Buffer) {
-    let separator = "─".repeat(area.width as usize);
-    Paragraph::new(separator)
-        .style(Style::default().add_modifier(Modifier::DIM))
-        .render(area, buf);
-}
-
-/// Renders keyboard hints
-pub fn render_hints(area: Rect, buf: &mut Buffer) {
-    let hints = vec![
-        Span::styled("q", Style::default().add_modifier(Modifier::BOLD)),
-        Span::raw(" / "),
-        Span::styled("Ctrl-C", Style::default().add_modifier(Modifier::BOLD)),
-        Span::styled(" exit", Style::default().add_modifier(Modifier::DIM)),
-    ];
-
-    Paragraph::new(Line::from(hints)).render(area, buf);
-}
 
 /// Renders the content area based on current state
 pub fn render_content(status: &InstallationStatus, area: Rect, buf: &mut Buffer) {
@@ -209,6 +39,7 @@ pub fn render_content(status: &InstallationStatus, area: Rect, buf: &mut Buffer)
     // 2. Final status (finished/failed)
     // 3. Progress (installation/configuration)
     // 4. Issues (blocking installation)
+
     // 5. Default stage message
 
     if !status.questions.is_empty() {
@@ -224,15 +55,6 @@ pub fn render_content(status: &InstallationStatus, area: Rect, buf: &mut Buffer)
     } else {
         render_stage(status, area, buf);
     }
-}
-
-/// Helper enum for busy state
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum BusyState {
-    Busy,
-    Idle,
-    Waiting,
-    Failed,
 }
 
 /// Renders questions
