+++ rust/agama-cli/src/monitor/ui/hints.rs
@@ -1,0 +1,41 @@
+// Copyright (c) [2024-2025] SUSE LLC
+//
+// All Rights Reserved.
+//
+// This program is free software; you can redistribute it and/or modify it
+// under the terms of the GNU General Public License as published by the Free
+// Software Foundation; either version 2 of the License, or (at your option)
+// any later version.
+//
+// This program is distributed in the hope that it will be useful, but WITHOUT
+// ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+// FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for
+// more details.
+//
+// You should have received a copy of the GNU General Public License along
+// with this program; if not, contact SUSE LLC.
+//
+// To contact SUSE LLC about this file by physical or electronic mail, you may
+// find current contact information at www.suse.com.
+
+//! Keyboard hints widget for the monitor TUI
+
+use ratatui::{
+    buffer::Buffer,
+    layout::Rect,
+    style::{Modifier, Style},
+    text::{Line, Span},
+    widgets::{Paragraph, Widget},
+};
+
+/// Renders keyboard hints
+pub fn render_hints(area: Rect, buf: &mut Buffer) {
+    let hints = vec![
+        Span::styled("q", Style::default().add_modifier(Modifier::BOLD)),
+        Span::raw(" / "),
+        Span::styled("Ctrl-C", Style::default().add_modifier(Modifier::BOLD)),
+        Span::styled(" exit", Style::default().add_modifier(Modifier::DIM)),
+    ];
+
+    Paragraph::new(Line::from(hints)).render(area, buf);
+}
