+++ rust/agama-cli/src/monitor/ui/mod.rs
@@ -1,0 +1,35 @@
+// Copyright (c) [2024-2025] SUSE LLC
+//
+// All Rights Reserved.
+//
+// This program is free software; you can redistribute it and/or modify it
+// under the terms of the GNU General Public License as published by the Free
+// Software Foundation; either version 2 of the License, or (at your option)
+// any later version.
+//
+// This program is distributed in the hope that it will be useful, but WITHOUT
+// ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+// FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for
+// more details.
+//
+// You should have received a copy of the GNU General Public License along
+// with this program; if not, contact SUSE LLC.
+//
+// To contact SUSE LLC about this file by physical or electronic mail, you may
+// find current contact information at www.suse.com.
+
+//! UI rendering modules for the monitor TUI
+
+mod content;
+mod hints;
+mod layout;
+mod product;
+mod separator;
+mod status_bar;
+
+pub use content::render_content;
+pub use hints::render_hints;
+pub use layout::create_layout;
+pub use product::render_product;
+pub use separator::render_separator;
+pub use status_bar::render_status_bar;
